% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicopula-package.R
\docType{package}
\name{bicopula-package}
\alias{bicopula}
\alias{bicopula-package}
\title{bicopula: Bivariate Copula Regression Models}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alessandro Gasparini \email{ellessenne@pm.me} (\href{https://orcid.org/0000-0002-8319-7624}{ORCID})

}
\keyword{internal}
